VueRouter: Fix incorrect parameters update for selected film route, move query parameters handling to StartPageHeader component, fix incorrect name for changeSearchTermActionPayload
