+++ src/components/search/SearchPanel.vue
@@ -33,14 +33,13 @@ import ButtonGroup from '@/components/common/ButtonGroup.vue';
 import PrimaryButton from '@/components/common/PrimaryButton.vue';
 import Vue from 'vue';
 import { FilmGetterProps } from '@/vuex/modules/films/getters';
-import { SearchByOptionNames, SearchQueryParams, SortByOptionsNames } from '@/enums/search';
+import { SearchByOptionNames, SearchQueryParams } from '@/enums/search';
 import { SearchByOptions } from '@/constants/search';
 import { SearchParams } from '@/types/search';
 import { areSearchParamsEqual, createSearchUrl, getSearchParams } from '@/utils/search';
 import {
-    changeSearchActionPayload,
     changeSearchByActionPayload,
-    changeSortByActionPayload,
+    changeSearchTermActionPayload,
     getFilmsRequestActionPayload,
 } from '@/vuex/modules/films/actions';
 import { getFilmModuleType } from '@/vuex/store/utils';
@@ -87,30 +86,8 @@ export default Vue.extend({
             this.$store.dispatch(changeSearchByActionPayload(option as SearchByOptionNames));
         },
         onChangeSearchQuery(value: string): void {
-            this.$store.dispatch(changeSearchActionPayload(value));
+            this.$store.dispatch(changeSearchTermActionPayload(value));
         },
-        onChangeSortByOption(option: string): void {
-            this.$store.dispatch(changeSortByActionPayload(option as SortByOptionsNames));
-        },
-    },
-    created(): void {
-        const searchQueryParams: SearchParams = getSearchParams(this.$route);
-
-        const searchTermQueryParameter: string | undefined = searchQueryParams[SearchQueryParams.SearchTerm];
-        const searchByQueryParameter: string = searchQueryParams[SearchQueryParams.SearchByOption];
-        const sortByQueryParameter: string = searchQueryParams[SearchQueryParams.SortByOption];
-
-        if (searchTermQueryParameter && searchTermQueryParameter !== this.searchTerm) {
-            this.onChangeSearchQuery(searchTermQueryParameter);
-        }
-
-        if (searchByQueryParameter && searchByQueryParameter !== this.searchByOption) {
-            this.onChangeSearchByOption(searchByQueryParameter);
-        }
-
-        if (sortByQueryParameter && sortByQueryParameter !== this.sortByOption) {
-            this.onChangeSortByOption(sortByQueryParameter);
-        }
     },
 });
 </script>
